+++ Sources/UI/Theme.swift
@@ -2,6 +2,4 @@ import Foundation
 
 enum Theme: Int {
     case system, light, dark
-    
-    
 }